backend/RAG Amna - Chatbot/requirements.txt
fastapi
uvicorn
cohere
python-dotenv
qdrant-client
